requirements.txt
agent
requests==2.26.0
PyYAML==5.4.1
numpy==1.21.2
pandas==1.3.3
openai==0.11.3